requirements.txt
Django
djangorestframework
websocket-client